First commit
Class_Maze_test is an adaptation of the maze created by Jan and Sid. I quickly updated it for the testing of my functions without keeping their generality for the next mazes.

functions contain the implementation of the 5 action selection algorithms for the 1st maze. They still need to be generalised for function approximations with neuro nets for maze 2 to 4