added second scoring function + NN
added second scoring function (steps are moves and not episodes) and my own implementation of a neuro net.
I didn't manage to use tensorflow on my computer, I had some weird bugs that may be due to the installation on my computer.

Good news is that with the new scoring function we seem to get the same results for the simple maze as in the paper

The weights of the neuro nets diverge very quickly, there may be an error on my implementation of the backpropagation (although I doubt it).

I didn't use any dropout nor biases term in the neuro net, maybe they did that for the paper but they don't mention it. I am not sure how to fix it...

I didn't implement the learning for the partially observable maze yet.

I didn't implement the learning for ACLA with the neuro net.